[Refactor] Change member variable types from double to float in Vehicle class for consistency and performance

+++ src/Headers/Physics/vehicle.h
@@ -14,21 +14,21 @@
  */
 class Vehicle {
 public:
-    double mass;   // Masse [kg]
-    double dist_cog_front_axle;   // Distance COG - essieu avant [m]
-    double dist_cog_rear_axle;   // Distance COG - essieu arrière [m]
-    double airRes;  // Coefficient de résistance de l'air
-    double vx;  // Vitesse longitudinale [m/s]
-    double vy;  // Vitesse latérale [m/s]
-    double r;   // Taux de lacet (vitesse angulaire) [rad/s]
-    double I;   // Moment d'inertie effectif pour le lacet
-    double Cx;  // Rigidité longitudinale [N]
-    double Cy;  // Rigidité latérale [N/rad]
+    float mass;   // Masse [kg]
+    float dist_cog_front_axle;   // Distance COG - essieu avant [m]
+    float dist_cog_rear_axle;   // Distance COG - essieu arrière [m]
+    float airRes;  // Coefficient de résistance de l'air
+    float vx;  // Vitesse longitudinale [m/s]
+    float vy;  // Vitesse latérale [m/s]
+    float r;   // Taux de lacet (vitesse angulaire) [rad/s]
+    float I;   // Moment d'inertie effectif pour le lacet
+    float Cx;  // Rigidité longitudinale [N]
+    float Cy;  // Rigidité latérale [N/rad]
 
     // Pour gérer la position du Véhicule
-    double x;
-    double y;
-    double psi;
+    float x;
+    float y;
+    float psi;
 
     /**
      * @param mass the mass of the vehicle
@@ -41,36 +41,35 @@ public:
      * @param y the y position of the vehicle
      * @param r the yaw rate of the vehicle
      */
-    Vehicle(const double mass,
-            const double distance_cog_front_axle, const double distance_cog_rear_axle,
-            const double airRes,
-            const double cx, const double cy,
-            const double x, const double y, const double r)
+    Vehicle(const float mass,
+            const float distance_cog_front_axle, const float distance_cog_rear_axle,
+            const float airRes,
+            const float cx, const float cy,
+            const float x, const float y, const float r)
     : mass(mass), dist_cog_front_axle(distance_cog_front_axle), dist_cog_rear_axle(distance_cog_rear_axle),
     airRes(airRes), vx(0.0), vy(0.0), r(r), Cx(cx), Cy(cy), x(x), y(y), psi(0.0) {
         I = mass * std::pow(0.5 * (distance_cog_front_axle + distance_cog_rear_axle), 2);
     }
 
 
-    void updateBicycle(const double dt, const double delta, const double slip)
+    void updateBicycle(const float dt, const float delta, const float slip)
     {
         // Calculer les angles de glissement pour les pneus avant (alpha_F) et arrière (alpha_R)
-        double alpha_F = 0.0, alpha_R = 0.0;
+        float alpha_F = 0.0, alpha_R = 0.0;
         if (vx > 0.01) { // évite la division par zéro
             alpha_F = delta - (vy + dist_cog_front_axle * r) / vx;
             alpha_R = -(vy - dist_cog_rear_axle * r) / vx;
         }
 
         // Calcul des forces sur les pneus (hypothèse : les forces sont identiques sur les deux roues de l'essieu)
-        const double F_x_front = 2.0 * Cx * slip;    // Force longitudinale sur l'essieu avant (drive wheels)
-        constexpr double F_x_rear  = 0.0;            // Pas de force longitudinale à l'arrière
-        const double F_y_front = 2.0 * Cy * alpha_F; // Force latérale sur l'essieu avant
-        const double F_y_rear  = 2.0 * Cy * alpha_R; // Force latérale sur l'essieu arrière
+        const float F_x_front = 2.0 * Cx * slip;    // Force longitudinale sur l'essieu avant (drive wheels)
+        constexpr float F_x_rear  = 0.0;            // Pas de force longitudinale à l'arrière
+        const float F_y_front = 2.0 * Cy * alpha_F; // Force latérale sur l'essieu avant
+        const float F_y_rear  = 2.0 * Cy * alpha_R; // Force latérale sur l'essieu arrière
 
-        // Calcul des accélérations selon le modèle "bicycle"
-        const double ax = vy * r + 1.0/mass * (F_x_front * cos(delta) - F_y_front * sin(delta) + F_x_rear - airRes * vx * vx);
-        const double ay = -vx * r + 1.0/mass * (F_x_front * sin(delta) + F_y_front * cos(delta) + F_y_rear);
-        const double r_dot = 1.0 / I * (dist_cog_front_axle * (F_x_front * sin(delta) + F_y_front * cos(delta)) - dist_cog_rear_axle * F_y_rear);
+        const float ax = vy * r + 1.0/mass * (F_x_front * cos(delta) - F_y_front * sin(delta) + F_x_rear - airRes * vx * vx);
+        const float ay = -vx * r + 1.0/mass * (F_x_front * sin(delta) + F_y_front * cos(delta) + F_y_rear);
+        const float r_dot = 1.0 / I * (dist_cog_front_axle * (F_x_front * sin(delta) + F_y_front * cos(delta)) - dist_cog_rear_axle * F_y_rear);
 
         // Mise à jour des états par intégration d'Euler
         vx += ax * dt;
@@ -80,8 +79,8 @@ public:
         psi += r * dt; // Integration du taux de lacet pour obtenir l'angle de direction
 
         // Transformation des vitesses locales en vitesses globales :
-        const double v_global_x = vx * cos(psi) - vy * sin(psi);
-        const double v_global_y = vx * sin(psi) + vy * cos(psi);
+        const float v_global_x = vx * cos(psi) - vy * sin(psi);
+        const float v_global_y = vx * sin(psi) + vy * cos(psi);
 
         // Mise à jour des positions globales par intégration :
         x += v_global_x * dt;
